Remove CngKeyLite
